Refactoring du scorer + fix + tests

+++ src/main/java/com/github/quadinsa5if/findingandqueryingtext/model/Entry.java
@@ -5,9 +5,9 @@ import java.util.Objects;
 public class Entry {
 
     public final int articleId;
-    public final double score;
+    public final float score;
 
-    public Entry(int articleId, double score) {
+    public Entry(int articleId, float score) {
         this.articleId = articleId;
         this.score = score;
     }
@@ -16,7 +16,7 @@ public class Entry {
         return articleId;
     }
 
-    public double score() {
+    public float score() {
         return score;
     }
 
@@ -33,7 +33,7 @@ public class Entry {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Entry entry = (Entry) o;
-        return Double.compare(entry.score, score) == 0 &&
+        return Float.compare(entry.score, score) == 0 &&
                 Objects.equals(articleId, entry.articleId);
     }
